requirements.txt
streamlit
langchain
groq
langchain-groq
python-dotenv
faster-wishper
gtts
soundevice
soundfile